refactor: deleteHistoriesTest parameterizedTest로 변경
